Moved psalter to the CLI command class
